refactor(ui): extract chat/useSessionStats hook from App.tsx
Moves token usage + context breakdown state, the /stats fetch with its
debounced variant, the while-generating 800ms poll, and the token baseline
ref into a hook. The streaming core still writes through the exposed setters
and baseline ref exactly as before. Pure move, no behavior change.
